Update Main.py
Redid some of the functionality, firstly to fix the bug where when a contact was updated it wouldn't open the view page again. Also made the program use the objects more rather than only using it to create an array, now the objects are the "main course" of the program.